Do away with durations
@gugu felt that having "days" as a construct would hamper UX grately.
And I agree too. It was one of the concerns I was having as well. The
simplest way to solve it, he said, is to just give a date input. Which
works amazingly well.

The mental model becomes:
- Select / create a sheet
- select a date
- enter expense

The system can then ascertain the duration of the sheet by itself :)